requirements.txt
pytesseract
opencv-python
pymupdf
pillow
resume-parser
nltk
spacy==2.3.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
scikit-learn
flask